refactor(st-09050): modularize tool builder
